Combine jest and tests files in integration tests
