+++ integration-tests/jest.config.js
@@ -2,7 +2,8 @@ module.exports = {
   verbose: false,
   preset: 'jest-puppeteer',
   testEnvironment: '<rootDir>/jest-circus-environment.js',
-  testRegex: '\\.integration-test\\.js$',
+  testMatch: ['<rootDir>/tests/*.js'],
+  testSequencer: '<rootDir>/test-sequencer.js',
   reporters: [
     'default',
     '<rootDir>/jest-reporter.js',
