--- integration-tests/jest/1-create-interactive-analysis.integration-test.js
@@ -1,5 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { testCreateInteractiveAnalysis } = require('../tests/create-interactive-analysis')
-
-
-registerTest(testCreateInteractiveAnalysis)
