--- integration-tests/jest/billing-projects.integration-test.js
@@ -1,8 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { testBillingMembers, testBillingSpendReport, testBillingWorkspaces, testDeleteBillingProject } = require('../tests/billing-projects')
-
-
-registerTest(testBillingMembers)
-registerTest(testBillingSpendReport)
-registerTest(testBillingWorkspaces)
-registerTest(testDeleteBillingProject)
