--- integration-tests/jest/find-workflow.integration-test.js
@@ -1,5 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { testFindWorkflow } = require('../tests/find-workflow')
-
-
-registerTest(testFindWorkflow)
