--- integration-tests/jest/janitor.integration-test.js
@@ -1,5 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { janitor } = require('../tests/janitor')
-
-
-registerTest(janitor)
