--- integration-tests/jest/preview-dataset.integration-test.js
@@ -1,5 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { testPreviewDataset } = require('../tests/preview-dataset')
-
-
-registerTest(testPreviewDataset)
