--- integration-tests/jest/preview-drs-uri.integration-test.js
@@ -1,5 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { testPreviewDrsUri } = require('../tests/preview-drs-uri')
-
-
-registerTest(testPreviewDrsUri)
