--- integration-tests/jest/register-user.integration-test.js
@@ -1,5 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { testRegisterUser } = require('../tests/register-user')
-
-
-registerTest(testRegisterUser)
