--- integration-tests/jest/run-catalog-filter.integration-test.js
@@ -1,5 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { testCatalogFilter } = require('../tests/run-catalog-filter')
-
-
-registerTest(testCatalogFilter)
