--- integration-tests/jest/run-catalog-workflow.integration-test.js
@@ -1,5 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { testCatalog } = require('../tests/run-catalog-workflow')
-
-
-registerTest(testCatalog)
