--- integration-tests/jest/workspace-dashboard.integration-test.js
@@ -1,6 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { googleWorkspaceDashboard, azureWorkspaceDashboard } = require('../tests/workspace-dashboard')
-
-
-registerTest(googleWorkspaceDashboard)
-registerTest(azureWorkspaceDashboard)
