+++ integration-tests/package.json
@@ -8,6 +8,7 @@
     "test-flakes": "FLAKES=true yarn test-local --runInBand"
   },
   "devDependencies": {
+    "@jest/test-sequencer": "^27.5.1",
     "jest": "^27.4.3",
     "jest-environment-puppeteer": "^6.0.3",
     "jest-junit": "^13.0.0",
