+++ integration-tests/tests/create-interactive-analysis.js
@@ -3,6 +3,7 @@ const { withRegisteredUser, withBilling, withWorkspace, performAnalysisTabSetup 
 const {
   click, clickable, getAnimatedDrawer, image, findElement, noSpinnersAfter, select, fillIn, input, findText
 } = require('../utils/integration-utils')
+const { registerTest } = require('../utils/jest-utils')
 
 
 const notebookName = 'analysis-test-notebook'
@@ -21,9 +22,8 @@ const testCreateInteractiveAnalysisFn = _.flow(
   await noSpinnersAfter(page, { action: () => click(page, clickable({ text: 'Create Analysis' })) })
   await findText(page, 'A cloud environment consists of application configuration, cloud compute and persistent disk(s).')
 })
-const testCreateInteractiveAnalysis = {
+
+registerTest({
   name: 'create-interactive-analysis',
   fn: testCreateInteractiveAnalysisFn
-}
-
-module.exports = { testCreateInteractiveAnalysis }
+})
