+++ integration-tests/tests/register-user.js
@@ -2,6 +2,7 @@ const { withUser } = require('../utils/integration-helpers')
 const { fillIn, findText, click, clickable, input, signIntoTerra } = require('../utils/integration-utils')
 const { fillInReplace } = require('../utils/integration-utils')
 const { waitUntilLoadedOrTimeout } = require('../utils/integration-utils')
+const { registerTest } = require('../utils/jest-utils')
 
 
 const testRegisterUserFn = withUser(async ({ page, testUrl, token }) => {
@@ -15,9 +16,7 @@ const testRegisterUserFn = withUser(async ({ page, testUrl, token }) => {
   await findText(page, 'To get started, Create a New Workspace')
 })
 
-const testRegisterUser = {
+registerTest({
   name: 'register-user',
   fn: testRegisterUserFn
-}
-
-module.exports = { testRegisterUser }
+})
